backend/requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
Levenshtein==0.26.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.0
pytz==2024.2
RapidFuzz==3.10.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
